Update all PHP dependencies
This way we can share the same dependencies across PHP 7.4 to PHP 8.1,
and simplify our CI pipeline

+++ tests/PhpWord/_includes/TestHelperDOCX.php
@@ -76,7 +76,7 @@ class TestHelperDOCX
      */
     public static function clear()
     {
-        if (file_exists(self::$file)) {
+        if (self::$file && file_exists(self::$file)) {
             unlink(self::$file);
         }
         if (is_dir(Settings::getTempDir() . '/PhpWord_Unit_Test/')) {
